Feat : VolumeAlcool And IsActive
gestion des propriété de l'alcool dans la table product (suppression de alcoholproduct)
adaptation du code de recherche sur le site web
désactivation d'un produit si conflit lors de la suppression avec d'autres tables
gestion de l'affichage d'un produit actif ou non dans la vue.

+++ NegoAPI/Services/ProductService/ProductService.cs
@@ -54,10 +54,19 @@ namespace NegoAPI.Services.ProductService
         public async Task<Product> DeleteProductAsync(Guid id)
         {
             var product = await GetProductByIdAsync(id);
-            if (product != null)
+            if (product == null) return null;
+            try
             {
-                _context.Products.Remove(product);
-                await _context.SaveChangesAsync();
+                if (product != null)
+                {
+                    _context.Products.Remove(product);
+                    await _context.SaveChangesAsync();
+                }
+            }
+            catch(DbUpdateException)
+            {
+                product.ProIsActive = false;
+                await UpdateProductAsync(product);
             }
             return product;
         }
